refactor: refactor and optimization phase 1
