+++ src/components/layouts/section.js
@@ -1,0 +1,10 @@
+/* eslint-disable */
+
+import React from "react";
+import CashRegisterManager from "../features/cash-register-manager";
+
+function Section() {
+  return <CashRegisterManager />;
+}
+
+export default Section;
