+++ src/config/app-config.js
@@ -1,4 +1,5 @@
 /* eslint-disable */
+
 const AVAILABLE_NOTES = [2000, 500, 100, 20, 10, 5, 1];
 
 const NO_OF_EACH_AVAILABLE_NOTES = [];
